libservo: Rename rust files to match the data structure they expose
